* CRI-pips: (41 commits)
  fix for ticket 349:
  Added 2nd level statement methods and name cleaning
  print line number when finding an error in sac config file
  some more cleaning in auto-*
  remove useless and non portable malloc.h
  Change name of gfc2pips parser (the binary was named f951 and is now gfc2pips)
  Fix Makefiles for Fortran95 : autotools doesn't like ?= and ifdef
  fix bugs introduced sooner
  Sorry Serge, it didn't compile !
  some more const in pips
  Allow to define GCC version in used for GFC2PIPS with an environment variable : PIPS_GFC2PIPS_GCC_VERSION
  Fix "compile" target
  Fix for autotools : make distcheck
  Fix a bug that broke compilation 
  fix typo
  Thats a huge commit :)
  fix a errant # in pipsmake-rc.tex that made property parsing fail, 
  Change representation of effects involving struct fields :
  Fix a bug, else was missing
  Fix autotools building 
  ...

+++ packages/PIPS/pips/src/Libs/preprocessor/csplit_file.c
@@ -190,7 +190,7 @@ void csplit_close_compilation_unit()
 */
 void copy_between_2_fd_up_to_offset(FILE * source,
 				    FILE * destination,
-				    size_t up_to_offset,
+				    unsigned long long up_to_offset,
 				    bool greedy_spaces) {
   int c = EOF;
   int next_c = EOF;
@@ -243,7 +243,7 @@ void copy_between_2_fd_up_to_offset(FILE * source,
 /* Copy the input file to the compilation unit between the function
    declarations up to the current function definition. */
 void csplit_append_to_compilation_unit(int last_line,
-				       size_t last_offset) {
+				       unsigned long long last_offset) {
   pips_debug(2, "append to compilation unit up-to line %d (from %d) or offset %zd\n",
 	     last_line, current_input_line, last_offset);
